feat: Add Shell Tree action
This commit introduces a GitHub Action called "Shell Tree" that generates a directory tree of the
repository.  The action takes an optional path input, defaulting to the root directory.  It
outputs a matrix, although the current implementation doesn't appear to generate a meaningful
matrix value.  The tree is generated using ls -R and formatted with sed.